modularize mul
