+++ tfjs-core/src/ops/ops.ts
@@ -64,6 +64,7 @@ export {maxPoolWithArgmax} from './max_pool_with_argmax';
 export {maximum} from './maximum';
 export {minimum} from './minimum';
 export {mod} from './mod';
+export {mul} from './mul';
 export {multinomial} from './multinomial';
 export {notEqual} from './not_equal';
 export {oneHot} from './one_hot';
