+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -54,6 +54,7 @@ const CHAINED_OPS = [
   'maxPool',
   'minimum',
   'mod',
+  'mul',
   'notEqual',
   'oneHot',
   'pad',
