+++ tfjs-core/src/register_all_gradients.ts
@@ -40,6 +40,7 @@ import {maxPool3DGradConfig} from './gradients/MaxPool3D_grad';
 import {maxPoolGradConfig} from './gradients/MaxPool_grad';
 import {minimumGradConfig} from './gradients/Minimum_grad';
 import {modGradConfig} from './gradients/Mod_grad';
+import {multiplyGradConfig} from './gradients/Multiply_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
 import {powGradConfig} from './gradients/Pow_grad';
@@ -87,6 +88,7 @@ const gradConfigs: GradConfig[] = [
   maxPool3DGradConfig,
   minimumGradConfig,
   modGradConfig,
+  multiplyGradConfig,
   oneHotGradConfig,
   padV2GradConfig,
   powGradConfig,
